refactor: 로또 당첨 결과를 출력하는 역할을 담당하는 별도 클래스 생성
